Fix parser error when reading large headers (#41)
* Fix parser error when reading large headers

The rosbag docs specify `conn_count` and `chunk_count` are Int32LE values, but we were reading them as Int16LE values.  This resulted in invalid parsing when either of those fields contained numbers larger than could be represented by an Int16LE.

Test plan: wrote unit tests to simulate larger bag headers and ensured they can be parsed correctly.

Semver impact: this is a semver __patch__ change - bugfix.

* Comments

+++ src/record.js
@@ -36,8 +36,8 @@ export class BagHeader extends Record {
   constructor(fields: { [key: string]: Buffer }) {
     super(fields);
     this.indexPosition = readUInt64LE(fields.index_pos);
-    this.connectionCount = fields.conn_count.readInt16LE(0);
-    this.chunkCount = fields.chunk_count.readInt16LE(0);
+    this.connectionCount = fields.conn_count.readInt32LE(0);
+    this.chunkCount = fields.chunk_count.readInt32LE(0);
   }
 }
